Factor out ClassUniquenessLockTask
