+++ gradle-baseline-java/src/main/groovy/com/palantir/baseline/plugins/BaselineClassUniquenessLockPlugin.java
@@ -16,23 +16,11 @@
 
 package com.palantir.baseline.plugins;
 
-import com.palantir.baseline.tasks.ClassUniquenessAnalyzer;
-import java.io.File;
-import java.util.Collection;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import org.gradle.api.Action;
-import org.gradle.api.GradleException;
+import com.palantir.baseline.tasks.ClassUniquenessLockTask;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.provider.SetProperty;
 import org.gradle.language.base.plugins.LifecycleBasePlugin;
-import org.gradle.util.GFileUtils;
 
 /**
  * This plugin is similar to https://github.com/nebula-plugins/gradle-lint-plugin/wiki/Duplicate-Classes-Rule but goes
@@ -51,87 +39,10 @@ public class BaselineClassUniquenessLockPlugin extends AbstractBaselinePlugin {
             configurationNames.add("runtimeClasspath");
         });
 
-        File lockFile = project.file("baseline-class-uniqueness.lock");
-
         // TODO(dfox): up-to-dateness of this task
-        Task checkClassUniquenessLock = project.getTasks().create("checkClassUniquenessLock");
-        checkClassUniquenessLock.doLast(new Action<Task>() {
-            @Override
-            public void execute(Task task) {
-                Map<String, Optional<String>> resultsByConfiguration = configurationNames.get().stream()
-                        .collect(Collectors.toMap(Function.identity(), configurationName -> {
-                            ClassUniquenessAnalyzer analyzer = new ClassUniquenessAnalyzer(project.getLogger());
-                            Configuration configuration = project.getConfigurations().getByName(configurationName);
-                            analyzer.analyzeConfiguration(configuration);
-                            Collection<Set<ModuleVersionIdentifier>> problemJars = analyzer.getDifferingProblemJars();
-
-                            if (problemJars.isEmpty()) {
-                                return Optional.empty();
-                            }
-
-                            StringBuilder stringBuilder = new StringBuilder();
-                            // TODO(dfox): ensure we iterate through problemJars in a stable order
-                            for (Set<ModuleVersionIdentifier> clashingJars : problemJars) {
-                                stringBuilder
-                                        .append(clashingJars.stream()
-                                                .map(mvi -> mvi.getGroup() + ":" + mvi.getName())
-                                                .sorted()
-                                                .collect(Collectors.joining(", ", "[", "]")))
-                                        .append('\n');
-
-                                analyzer.getDifferingSharedClassesInProblemJars(clashingJars).stream()
-                                        .sorted()
-                                        .forEach(className -> {
-                                            stringBuilder.append("  - ");
-                                            stringBuilder.append(className);
-                                            stringBuilder.append('\n');
-                                        });
-                            }
-                            return Optional.of(stringBuilder.toString());
-                        }));
-
-                boolean conflictsFound = resultsByConfiguration.values().stream().anyMatch(Optional::isPresent);
-                if (!conflictsFound) {
-                    ensureLockfileDoesNotExist();
-                } else {
-                    StringBuilder stringBuilder = new StringBuilder();
-                    stringBuilder.append(
-                            "# Run ./gradlew checkClassUniquenessLock --write-locks to update this " + "file\n\n");
-                    resultsByConfiguration.forEach((name, contents) -> {
-                        if (contents.isPresent()) {
-                            stringBuilder.append("## ").append(name).append("\n");
-                            stringBuilder.append(contents.get());
-                        }
-                    });
-                    ensureLockfileContains(stringBuilder.toString());
-                }
-            }
-
-            private void ensureLockfileContains(String expected) {
-                if (project.getGradle().getStartParameter().isWriteDependencyLocks()) {
-                    GFileUtils.writeFile(expected, lockFile);
-                } else {
-                    String onDisk = GFileUtils.readFile(lockFile);
-                    if (!onDisk.equals(expected)) {
-                        throw new GradleException(lockFile
-                                + " is out of date, please run `./gradlew "
-                                + "checkClassUniquenessLock --write-locks` to update this file");
-                    }
-                }
-            }
-
-            private void ensureLockfileDoesNotExist() {
-                if (project.getGradle().getStartParameter().isWriteDependencyLocks()) {
-                    GFileUtils.deleteQuietly(lockFile);
-                } else {
-                    if (!lockFile.exists()) {
-                        throw new GradleException(lockFile + " should not exist (as no problems were found).");
-                    }
-                }
-            }
-        });
+        Task task = project.getTasks().create("checkClassUniquenessLock", ClassUniquenessLockTask.class);
 
         project.getPlugins().apply(LifecycleBasePlugin.class);
-        project.getTasks().getByName(LifecycleBasePlugin.CHECK_TASK_NAME).dependsOn(checkClassUniquenessLock);
+        project.getTasks().getByName(LifecycleBasePlugin.CHECK_TASK_NAME).dependsOn(task);
     }
 }
